+++ cocos/platform/CCImage.cpp
@@ -982,7 +982,7 @@ bool Image::initWithJpgData(uint8_t * data, ssize_t dataLen)
     struct MyErrorMgr jerr;
     /* libjpeg data structure for storing one row, that is, scanline of an image */
     JSAMPROW row_pointer[1] = {0};
-    unsigned long location = 0;
+    uint32_t location = 0;
 
     bool ret = false;
     do 
@@ -2319,7 +2319,7 @@ bool Image::saveImageToJPG(const std::string& filePath)
         CC_BREAK_IF(nullptr == outfile);
 
         unsigned char* outputBuffer = nullptr;
-        unsigned long outputSize = 0;
+        uint32_t outputSize = 0;
         jpeg_mem_dest(&cinfo, &outputBuffer, &outputSize);
 
         cinfo.image_width = _width;    /* image width and height, in pixels */
