+++ cocos/platform/win32/CCStdC-win32.cpp
@@ -36,8 +36,8 @@ int gettimeofday(struct timeval * val, struct timezone *)
         LARGE_INTEGER liTime, liFreq;
         QueryPerformanceFrequency( &liFreq );
         QueryPerformanceCounter( &liTime );
-        val->tv_sec     = (long)( liTime.QuadPart / liFreq.QuadPart );
-        val->tv_usec    = (long)( liTime.QuadPart * 1000000.0 / liFreq.QuadPart - val->tv_sec * 1000000.0 );
+        val->tv_sec     = (int32_t)( liTime.QuadPart / liFreq.QuadPart );
+        val->tv_usec    = (int32_t)( liTime.QuadPart * 1000000.0 / liFreq.QuadPart - val->tv_sec * 1000000.0 );
     }
     return 0;
 }
